Significantly broaden MyProfile iOS SDK sample
Summary:
Based on Doug's recent feedback, we decided to broaden the MyProfile SDK sample to be more of a Hello World type sample.

The changes introduced are intended to have minimal UI componentry, so as to focus on the Facebook-ness of them.

Meaningful changes:
- Have 3 buttons in the UI for posting a status update, posting a photo, and invoking the friend picker.
- The status update just creates a new status with the logged in user's name and the time.  (And we need to ask for "update_status" permission in the session creation in order to support this.)
- The photo upload just uploads the app icon that we already have in the application.
- The friend picker inclusion also warranted turning the app into UINavigationController based one.

The results of all of these are just displayed in modal UIAlertViews.

iPhone and iPad nibs have been updated.  The code is pretty well commented.

I also modified the FB app on the server to have default privacy of "Only Me", which is definitely appropriate for this sample.

Note that I haven't actually renamed this HelloFacebook because a) it's a pain to do; and b) I didn't want these changes to get lost in a larger change.  I can do it as a follow on, but I don't think it's bad if we just keep it named MyProfile, and elevate it in the docs, etc., as a more comprehensive example.

Test Plan: My machine, lots of install/uninstall cycles, iPhone and iPad.

Reviewers: clang, jacl, ekoneil

Reviewed By: jacl

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D508038