[ios-sdk] Clean up samples to use new FBSession.activeSession functionality.
Summary:
Cleaned up samples that are not specifically about session management to use the static
FBSession.activeSession functionality to manage session state.

Also made some minor fixes:
- Renamed classes/files in PlacePickerSample and ProfilePictureSample to reflect
iOS conventions.
- Fixed BooleanOGSample to display friend activity in the case where they have
no "OR" actions.

Scrumptious is not included in this commit, as it depends on an FBSession change that
is still in review.

Test Plan:
- Ran each sample

Revert Plan:

Reviewers: jacl, mmarucheck, ayden, gregschechte

Reviewed By: jacl

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D510167