[ios-sdk] Changes to README, install package
Summary:
Added a README file, identical to the README.mdown, but less scary to people who don't know what an .mdown file is.
Added basic Getting Started steps to the README, and redirecting to the web Getting Started for richer docs.
Displaying the README on the final page of the Installer as well as the first one, to give developers another
change to read it before dismissing the installer.

Test Plan:
- Ran the installer

Revert Plan:

Reviewers: jacl, mmarucheck, ayden, gregschechte

Reviewed By: gregschechte

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D510247