Simplify "me" retrieval in MyProfile sample
Summary:
Simplifying MyProfile sample to:

- remove the storage of a FBRequestConnection since we never use it.
- use the [FBRequest startWithSession:graphPath] rather than a two-step deal.

With that simplified, what I still find the most confusing about this sample is why the "cached token" case needs to be dealt with explicitly and why we need to re-issue the session opening in that case.  Why isn't it done implicitly and this state FBSessionStateCreatedTokenLoaded hidden from the developer?  This next bit of code seems very unintuitive:

        // because we use this method to bootstrap button setup, we may find that we also need
        // to open a session that is ready and waiting with a cached token
        if (FBSession.activeSession.state == FBSessionStateCreatedTokenLoaded) {
            [FBSession.activeSession openWithCompletionHandler:^(FBSession *session,
                                                                 FBSessionState status,
                                                                 NSError *error) {
                [self updateForSessionChange];
            }];
        }

Test Plan: My machine.

Reviewers: jacl, clang, ekoneil

Reviewed By: jacl

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D508085