Update FBLoginView to include a button with Login/Logout
Summary:
Add a simple button with the Login/Logout text to the FBLoginView.  This lets us remove text in samples that use this.

Note that this presentation of the FBLoginView will change post-beta to be much more FB-branded, and this is just intended to tide us over.

Test Plan: MyProfile sample on iPhone/iPad simulators

Reviewers: jacl, clang

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D513372