Updated FBProfilePictureView to use hi-res square image Graph API
Summary:
The Graph API for retrieving hi-res square profile pictures is now on, and this change modifies the FBProfilePictureView to use it.

- For the logic change here, see FBProfilePictureView.m
- Changed the default cropping to square, since with hi-res square, the quality is typically just as good, and non-square is very hard to reason about as someone using the control (aspect ratio is different for every user).
- Removed all but the largest non-square blank profile, and updated the square one to be bigger.  No need to carry the smaller versions as these larger ones scale down fine.
- Updated ProfilePictureSample to include a "none" option that results in showing the blank profile.
- Updated HelloFacebookSample to make both profile picture views square.

Note that when you use this sample or control, some of the square images will still be 50x50.  That's because there's a migration job starting today (7/10) that should take about a week to re-gen larger square profiles for users who haven't updated their profile in more than a year.

Task: #1113853

Test Plan: Samples on iPhone/iPad.  In particular used ProfilePictureSample to run through lots of combinations.

Reviewers: clang, jacl, ayden, mmarucheck

Reviewed By: clang

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D515828