[ios-sdk] Fix bug in paging loader that led to multiple activity indicators in friend picker.
Summary:
FBGraphObjectPagingLoader was setting a dataNeededDelegate on its data source even if it was not in the
AsNeeded paging mode. This was fooling the data source into thinking it should display a spinner in the
bottom-most table cell to indicate data was being fetched. Controls that aren't in AsNeeded mode may have
other UI to indicate when they load data (as friend picker does), so this spinner was redundant.

Fixed FBGraphObjectPagingLoader to only set a dataNeededDelegate if it actually should; in turn, this pointed
out that pagingMode should not be a writable property on the object, as there is little hope of things working
correctly if the pagingMode is changed during the lifetime of an object. Changed it into a readonly property
set at init-time.

Test Plan:
- Ran FriendPickerSample
- Ran PlacesPickerSample
- Ran Scrumptious
- Ran HelloFacebook

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden

Reviewed By: jacl

CC: msdkexp@, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D516326