[ios-sdk] Removed string parameter from [FBSessionTokenCachingStrategy clearToken]
Summary:
The string parameter to this method was intended as an optimization to help implementations find
the token to be deleted. However, since the intention is that only the current token would ever be
deleted, they can store this data internally if they need to optimize for it, without requiring
that all callers pay the tax of passing it in.

Test Plan:
- Ran unit tests
- Ran SwitchUserSample (the only sample that used this method explicitly)

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden

Reviewed By: jacl

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D520450

Task ID: 1188716