Adds high-level FBMyData facebook integration object
Summary:
The FBMyData facebook integration object is meant to provide a super
lightweight integration experience for a core set of common needs. Apps
that have heavier weight needs may choose to use FBMyData in addition to
the more fundamental classes such as FBRequest and FBSession. However,
the goal of this class is to provide sufficient functionality that
some classes of application can get away with use of this class alone.
The capabilities are:
* notification of user logged in status
* ability to handle (but not present) UX for connecting/disconnecting
* ability to post a status
* ability to post a photo
* ability to fetch your info (name, fbid, etc.)
* ability to fetch your friends list

This diff also introduces an ultra-lightweight sample called "HiFacebook"

Test Plan: Ran through supported features using HiFacebook

Reviewers: clang, ayden

Reviewed By: clang

CC: msdkexp@, yariv, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D524684

Task ID: 1188698