[ios-sdk] Add FBUserSettingsViewController and FBViewController classes; pickers derive from FBViewController.
Summary:
FBViewController adds functionality for views which are presented modally, providing optional Done and Cancel
buttons in a title toolbar. Views presented in a navigation controller may have a Done button, but no
Cancel button (since the "back" button serves that purpose).

FBUserSettingsViewController displays the current logged-in state of the user, and allows the user to log out
and log in. The UI for this view controller is not yet final; in particular, it does not look good on iPad.
Waiting for feedback from design before finalizing.

Scrumptious uses FBUserSettingsViewController to let the user log out, but continues to use its custom view for
logging in.

Test Plan:
- Ran Scrumptious
- Ran new test app (not checked in) that exercises the view controllers in both a modal and a navigation context.

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden

Reviewed By: jacl

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D525096

Task ID: 1188684