Fixed a potential crash after loading a feed dialog in some cases
Following the instructions/code on https://developers.facebook.com/docs/howtos/feed-dialog-using-ios-sdk/

If ARC is enabled (but weak reference isn't enabled), by the end of the method, the facebook object is deallocated. When the FBRequestConnection is calling completeDeprecatedWithData, the app crashes because the variable delegate refers to a zombie Facebook object.

Anyway it's a good practice to set all unretained pointers that references self to nil during dealloc.