Removes preview iOS6 support, for RTM
Summary:
iOS6 integration in the codebase is provisional and for preview only,
until iOS6 ships. Because we will soon be merging the codebase to master
for use in production, we remove the support. Preview branches will
continue to have and update the iOS6 preview until the OS is released.

Test Plan: reran unit tests, ran a few samples

Reviewers: clang, gregschechte, ayden

Reviewed By: gregschechte

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D538958

Task ID: 1235919