Improves developer model for error handling in FBRequestConnection
Summary:
This changes requires a rational; here goes. The fix is that now the userInfo of the
error object works the same whether or not the request is a batch. This significantly
improves understandability, as well as makes it more feasible to write apps that dynamically
choose to batch or not. That's the good news. There are actually two reasons to make this
change now:
1) The previously mentioned improvment to the developer model
2) Changing this aspect of the developer model later would be a particularly painful
class of breaking change, that would likely introduce difficult to find/repro bugs in apps

The tradeoff is that the elegant fix to make improve the developer model is also a deep
change that would be too risky given that we are about to move these bits to production. So
we defer that fix until we have a time-window to stablize the change, and we make a less
elegant, but more stable fix here -- at a skin-deep point in the code.

Test Plan: ran unit tests, ran impacted samples

Reviewers: gregschechte, clang, mmarucheck

Reviewed By: mmarucheck

CC: msdkexp@, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D538988

Task ID: 1256476