[ios-sdk] Fix crash in FBGraphObjectSelection
Summary:
A previous commit added functionality to [FBGraphObjectDataSource clearGraphObjects] that actually
cleared the exising graph objects (as the name implies). However, this method was also being called
whenever a new request is issued, and clearing graph objects is not what we want to do in that case,
since during the period between the request being issued and results being returned, the user is still
free to interact with the previously-received (or cached) data.

Split clearGraphObjects into two methods and called the new method in the new-request scenario.

Test Plan:
- Ran Scrumptious
  - Selected cached friends in friend picker before and during request processing, no crash

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden, liangz

Reviewed By: jacl

CC: msdkexp@, platform-diffs@lists

Differential Revision: https://phabricator.fb.com/D540026

Task ID: 1318160