[ios-sdk] Fix bug where place picker cache descriptor was ignoring searchText; fix Scrumptious to call it correctly.
Summary:
Two bugs conspired to make caching inconsistent in Scrumptious. Scrumptious was passing "restaurant" as the search text
for the place picker, but not passing it when creating a cache descriptor for background loading. The cache descriptor itself
was also not properly using the searchText to build its URL. Fixed both of those.

Also fixed Scrumptious to be less strict in the accuracy it wants, and don't fill the cache from intermediate low-accuracy
location updates to avoid unnecessary network traffic.

Test Plan:
- Ran unit tests
- Ran Scrumptious on device, did not notice delay when displaying places as I did without this fix

Revert Plan:

Reviewers: jacl, mmarucheck, gregschechte, ayden

Reviewed By: jacl

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D540969