Clear cached requests and pictures when session is closeAndClear-ed
Summary:
When FBSession.closeAndClearTokenInformation() is called, we can go ahead and clear out some of the cached data.
Cached requests will have the session's accessToken in the url query params - which can be used as a filter
Stuff from akamai (images) does not have the same. So we just clear all to err on the side of safety w.r.t. privacy

Test Plan:
Manual testing of Scrumptious with lots of NSLogging.
Multiple login/logouts with different users.

Reviewers: jacl, clang

Reviewed By: jacl

CC: msdkexp@

Differential Revision: https://phabricator.fb.com/D551216

Task ID: 1129032