acquire the iOS6 advertisingIdentifier, if not opted out, and send it during publish.
Summary: Apple has added a new API in iOS6, and we should use it when available.

Test Plan:
run a publish in all 4 scenarios of iOS5/6 device X all sample applications.
Verify the the build never fails.
make sure to delete the apps on the phone before/after each test.
Only on ios6 should the advertiser_id value be set when publishing to facebook.

Reviewers: jacl, clang

Reviewed By: jacl

CC: vijaye, teck, calvin, msdkexp@, neko-dev@, suhas

Differential Revision: https://phabricator.fb.com/D574669

Conflicts:
	src/facebook-ios-sdk.xcodeproj/project.pbxproj