Don't display a dialog if the user didn't pick a place from the place picker.
Summary: Don't display a dialog if the user didn't pick a place.

Test Plan:
Click done in the place picker both with a place picked and without to see
if any problems crop up.

Click cancel in the place picker to make sure that it still dismisses the
dialog.

Reviewers: jacl

Reviewed By: jacl

CC: mmachedon

Differential Revision: https://phabricator.fb.com/D576940

Task ID: 1407166