Bucketizes the permissions developer model by read and publish
Summary:
This change establishes the developer model that supports the direction of
dissallowing permission requests to users that present both read and publish
permissions in a single request gesture. The specific changes are:
* Adds read and publish versions of permissions properties to certain controls
* Adds read and publish versions of open and reauthorize methods on FBSession
* Deprecates methods and properties that take unspecified collections of permissions
* Updates samples

Note - in order to unblock docs work, we are pushing this with a known build break in
one of the samples. Will fix the sample shortly.

Test Plan: Tested impacted scenarios via samples

Reviewers: clang, mmarucheck, gregschechte

Reviewed By: mmarucheck

CC: chrisp, ekoneil

Differential Revision: https://phabricator.fb.com/D580068

Task ID: 1731992