Added a login behavior for attempting the system login without falling back to other behaviors
Useful for application actions that could use publish permissions if available, but don't require them, and don't want to bother the user if publish hasn't already been authorized. On iOS5, the reauthorize flow is rather heavy and jarring and not at all suitable for optional permissions. Instead, let the application handle prompting the user for optional permissions when the system account isn't available (ie. iOS5).