Correctly invoke handler on -[FBURLConnection cancel]
Problem:
Caller's completion block is not called on a FBURLConnection when that
connection is terminated by -cancel such as due to a timeout.

Cause:
-cancel sets self.handler to nil before calling invokeHandler. The
cancel method retains a copy of the handler before calling
invokeHandler.

While invokeHandler uses the handler parameter for calling, it was
checking self.handler for non-nil before doing so.

All other calls to invokeHandler are made with self.handler as the
parameter so the issue wasn't apparent in 'normal' use.