Fix infinite notification loop in LoginView.
[FBLoginView wireViewForSession:] would be called by the global notification of
a new active session.  However, if the session state was
FBSessionStateCreatedTokenLoaded, it would call [FBSession
openActiveSession...], which would create and set a *new* activesession,
creating a new notification, etc.

Instead, in that case we use openWithBehaviour directly on that session rather
than forcing a new one to be created; I believe this was the intended
behaviour.