Updating Submodules Schemes
carthage update on any Facebook-SDK for iOS Development causes a timeout when checking out the dependancies for facebook-ios-sdk

Many companies restrict using the git:// scheme due to it using a specific port and not using proper authorization. Https is much more secure and more commonly unblocked.

Proposing this file be standardized to use https instead of both git and https schemes.

This fixes the carthage update timeout issue.

Alternatively users cloning this project will need to run a command similar to:

git config --global url."https://".insteadOf git://

or remove --global and do it on a project-to-project basis.