Correct translation string formatting
There were a couple of errors in the string formatting in `lrpoj`
translation string files which were causing issues when compiling an app
which uses the SDK.

Prior to this change, the app compile would fail with a JSON error
`Property List error: Unexpected character / at line 1`. With these
changes the app successfully compiles.