Escaped method_args to fix video sharing
Summary:
The method argument for sharing a video contains an embedded assetURL value that has an ampersand:
  method_args = {
    "video": {
      "assetURL":"assets-library:\/\/asset\/asset.mp4?id=9EF48B86-A9FF-4136-8009-5170327A2183&ext=mp4"
    },
    "shareUUID":"B829D2F7-C990-440A-BB43-D4C2F0FF2995",
    "dataFailuresFatal":true
  }

The SDK constructs a bridge URL that includes the following arguments in the query string:
  bridge_args=
  method_args=
  version=

The dictionaryWithQueryString: method used to construct the bridge URL tokenizes the query string using an ampersand delimiter. This unsophisticated approach ended up breaking the method_args argument in half. The main app ended up receiving the following:
  bridge_args=
  method_args=
  ext=
  version=

The method_args parameter was incomplete and couldn't be parsed as valid JSON, thus video sharing via SDK was not working.

This diff escapes the ampersand within the method_args parameter as %26 so the assetURL is still functional but does not get split apart when constructing the bridge URL.

Reviewed By: sachin004

Differential Revision: D12917676

fbshipit-source-id: bb3aab13ccd3204afd10a7627ec149aad6a80393