Support sharing video asset or data to ShareSheet
Summary:
Trying to share a video asset or video data via the ShareSheet wasn't working. Validation wasn't in place to prevent this and the underlying data was inaccessible since the iOS Share Extension supports sharing images and URLs.

This diff adds support for sharing a video asset by obtaining the legacy assets-library URL and for sharing video data by saving the video to a temporary file and sharing the URL to that local file.

Reviewed By: codytwinton

Differential Revision: D13032984

fbshipit-source-id: aa7782b4d2dc2a695158c4e4dc1fa898426df8dd