improve codeless RN support in ios
Summary:
our previous button detection and event fire methods are very tricky and miss some corner cases. For example, if a clickable view has two components as the test plan shows. When we click the image button, the react tag comes from the clicked image. And in our previous method we only check the first subview of the view and that won't fire events when users click other subviews.

we can improve by:
1. button detection: check view's all subviews and call reactTagAtPoint to check whether there is one mathced react tag, if so, then the view is a button.

2. event fire:  reactBinding here will be button's reactTag => eventBinding. If there is a 'touchEnd' event, we will find the react Tag of targetView's super button and then fire events.

Reviewed By: codytwinton

Differential Revision: D13339199

fbshipit-source-id: 7555774ce87c6b8395fbe8efba3e2ff6814a6223