Fix plus sign URL decoding
Summary:
Challenges that contain a blank space (`%20`) are causing login to fail with a 308 error.

This is because we are comparing an encoding of the challenge with a plus sign in place of the space. The URLDecode method in FBSDKUtility attempts to replace plus signs with a space, but it does this before the percent encoded characters are decoded, so it ends up doing nothing.

Reviewed By: codytwinton

Differential Revision: D13522544

fbshipit-source-id: a4ebbf603494db43e3e847cc9eb07797627923b6