Adjust web view for safeAreaInsets
Summary: The webview login option is broken on iPhone X and other newer iPhones with rounded corners and a notch. The 'X' close button is cut off by the rounded corner and so is not usable. Additionally, the padding below the webview should be expanded to allow the proper space for the home bar.

Reviewed By: codytwinton

Differential Revision: D13328432

fbshipit-source-id: 5af026af7efcc06703527862bba22497fa79b875