Replace NSURL alloc/initWithString with URLWithString for constants
Summary:
Replacing calls to [[NSURL alloc] initWithString:] that take constant values to [NSURL URLWithString:].

The Apple documentation for initWithString: states that the URLString parameter must not be nil. In practice it causes an exception when that's the case.

The Apple documentation for URLWithString: states that if URLString is malformed or nil, the class method returns nil.

Reviewed By: kolinkrewinkel

Differential Revision: D13541042

fbshipit-source-id: 74ae52de8b8513aa0d3992c1e48f627a4d2ab489