Move share extension code to a separate file
Summary:
This diff moves SDK code specific to the Share Extension to a separate file so that it can be used by both the current FBSDKShareKit library and the future lightweight FBSDKPlatformShareKit library.

Constants were added for strings that need to be accessed both when creating the parameter payload (in the SDK) and parsing the payload (in the Share Extension and main app, which will eventually link with the new lightweight library so these are declared in one place).

The code remaining in FBSDKShareDialog.m was also cleaned up with const, generics, some sanity checks and a bit of renaming for clarity.

Reviewed By: robtimp

Differential Revision: D13936761

fbshipit-source-id: 53102fde5d6f494548ea7b4d1bdf0f508c289511