Don't dismiss Safari view when returning from background
Summary:
Backgrounding the app and returning from background on iOS 11 and 12 causes the SafariViewController to be dismissed immediately. This does not happen on iOS 10 and earlier because we manually create the SafariViewController and store it in a property. However, on iOS 11 and 12 we use SFAuthenticationSession and ASWebAuthenticationSession, respectively. These classes create their own SafariViewControllers, so we should add a check if the topMostViewController is an instance of the class SFSafariViewController and store it in the `_safariViewController` instance variable when appropriate.

This will allow the behavior of iOS 11 and 12 to match earlier versions when returning from background, and is especially important when a code generator is used (since the user will need to switch apps and return).

Reviewed By: codytwinton

Differential Revision: D14027086

fbshipit-source-id: 7fdcd8f5ba905a1d1cef915126a16731c4cfdcc7