Ignore focused xcode project
Summary:
This should ensure that the focused project for the ios-sdk is always ignored. It seems like there was already a higher level ignore for focused projects in the fbobjc .gitignore but it seems like it wasn't applied in the specific case when this bug was reported. Hopefully this acts as a second layer of coverage.

Also, was not able to reproduce the bug so it may be wise to not make this change and wait until we have better repro steps.

Reviewed By: codytwinton

Differential Revision: D14087691

fbshipit-source-id: dd5c65afd350d79140c0449718d6cd6f84f62835