Fix FBSDKLoginKitTests
Summary:
This diff accomplishes a couple things:

**Fake Challenge**

Replaces the fake challenge in the unit tests with a new challenge that does not have a plus sign. Did this because currently there is a known issue due to the following:

Given there is one or more plus signs in the string, we use `FBSDKUtility` to UTF8 encode the string, leaving the plus signs unmodified. Prior to decoding we use `FBSDKUtility` to replace any plus signs in the UTF8 encoded string with whitespace.

As part of the login completion, we replace plus signs from the decoded string with whitespace. Therefore it is likely that there are bugs.

**Old Logic**

Removed tests that were related to old logic that represented a lack of change in permissions as a cancellation of login.

This was mixing concerns and was confusing to developers. It was fixed in D8676310.

Reviewed By: robtimp

Differential Revision: D14184211

fbshipit-source-id: 30437eb22243fe67d61774df158e97eca6bca826