Use Your Own FB App ID in RPS Sample
Summary: This diff is to update the RPSSample app to allow users to use their own FB App ID in the app. The previous version had an FB App ID in the plist file which limited viewing the events to only authorized users. I removed the FB App ID and added code to display an alert for the user to add their own FB App ID. I also revised the ReadMe.txt to include instructions on this change.

Reviewed By: joesus

Differential Revision: D14328680

fbshipit-source-id: e5dbd9c8f5c25e55d2a80f4f5821b3b600a474e3