Replaced NSURL alloc/initWithString: in SDK-related files
Summary:
As per the task, calling [[NSURL alloc] initWithString:] with a nil value will cause an exception. Calling [NSURL URLWithString:] with a nil value will fail gracefully by returning a nil value.

This diff Replaces calls to alloc/initWithString with URLWithString to mitigate the risk of crashing in SDK-related files.

Reviewed By: bsoyluoglu

Differential Revision: D14934299

fbshipit-source-id: c7dd5481d991044d240a14d0e2ee079e83613439