Change Xcode version to 10.1 for github release
Summary:
There is bitcode version change introduced by Xcode 10.2, so the frameworks we generated via Xcode 10.2 won't be able to be used in Xcode 10.1.

To support manual integration with Xcode 10.1, we need to build the frameworks with Xcode 10.1.

FYI: frameworks built via Xcode 10.2 could be used in newer version of Xcode (e.g. Xcode 10.2)

Reviewed By: codytwinton

Differential Revision: D15351874

fbshipit-source-id: 97f9f6babd2b1a4b5b1e634dd1f49e4b9d837321