Mark FBSDKShareOpenGraphContent as deprecated
Summary:
This diff is part of the Open Graph SDK deprecation effort.

There are two flavors of Open Graph shares: those without an Object URL (they supported custom actions and images but had no associated URL) and those with an Object URL (they also supported an optional hashtag).

Open Graph shares without an Object URL are being completely deprecated. Attempting to share such content via the Native app will return an error while the Share Extension will open an empty composer.

Open Graph shares with an Object URL will continue to be supported for apps using older versions of the SDK; however, they should be converted to ordinary link shares.

This diff adds the deprecated message to the FBSDKShareOpenGraphContent object so developers see a warning when building with this and are encouraged to move to FBSDKShareLinkContent instead.

Reviewed By: mirko02

Differential Revision: D15725629

fbshipit-source-id: 6b59bcef04c00a01f77e509aea429023ff1f0fa0