Removing Cryptography from ObjC SDK
Summary:
The encryption between the ObjC SDK and FBPlatformCoreKit is not secure. Since the cipher key is passed unencrypted via a URL it is easy to spoof or otherwise tinker with.

It is better to remove this so that other devs do not think that this is a safe way of passing information between installed apps.

Differential Revision: D15943543

fbshipit-source-id: 13da1fdb8b7e37df4ae57c00d42f0298764cbac1