Removing Cryptography from ObjC SDK (#996)
Summary:
Pull Request resolved: https://github.com/facebook/facebook-objc-sdk/pull/996

The encryption between the ObjC SDK and FBPlatformCoreKit is not secure. Since the cipher key is passed unencrypted via a URL it is easy to spoof or otherwise tinker with.

It is better to remove this so that other devs do not think that this is a safe way of passing information between installed apps.

Differential Revision: D15943543

fbshipit-source-id: 77597e6451dd595b8ac8607f9e8767e92b33f6c0