Fix FB Login on iOS 13 beta for static library release
Summary: We used compile time flags to detect whether iOS 13 is available, this woks for cocoapods as apps will compile the SDK source code, however, we also provide static and dynamic libraries, which won't have the fix. This is to fix the issue in compiled SDK instead of source code.

Reviewed By: tianqibt

Differential Revision: D16902381

fbshipit-source-id: ab29a0f77778d4c6f5ac18e65ab4758d90cd687b