Cleanup: There is only one login behavior available (#1027)
Summary:
Thanks for proposing a pull request.

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] describe the change (for example, what happens before the change, and after the change)

Deprecation of `FBSDKLoginBehavior`. All login flows utilize the browser as deep-linking directly to Facebook is discouraged. Native login is still possible but will happen via a step where a browser is invoked first. The ability to pass the option is present on `FBSDKLoginManager` and `FBSDKLoginButton` but the property is ignored. This marks that property as deprecated on both.
Pull Request resolved: https://github.com/facebook/facebook-objc-sdk/pull/1027

Reviewed By: jingping2015

Differential Revision: D17120143

Pulled By: joesus

fbshipit-source-id: 1fc30aac938b7387d7655e74a46d2d1333f85a3d