Separate Swift targets (#1031)
Summary:
Thanks for proposing a pull request.

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] describe the change (for example, what happens before the change, and after the change)

This is a step towards aligning Swift and Objective-C into a single SDK.

Currently we cannot ship a single target that includes both Swift and Objective-C because some consumers of the SDK are tied to older versions of Xcode. Including Swift would increase their binary size or force them to upgrade before Apple mandates an upgrade.

What we can do is prepare for this inevitability by moving the Swift files into this repository.

Additionally this is easier to maintain and allows us to us to reference ObjC files from Swift directly instead of treating them as a third party dependency.

This work raises a few questions about distribution channels (what will the starting versions be for instance) but that is an acceptable cost for the benefits we get from having all of our related code in one place.
Pull Request resolved: https://github.com/facebook/facebook-objc-sdk/pull/1031

Test Plan:
Able to build is the primary concern. That the old targets are not impacted by the new targets and that the new targets build correctly.

Ensure the older targets can build on Xcode 10.1
* FBSDKCoreKit
* FBSDKShareKit
* FBSDKLoginKit

Ensure the older AND newer targets can build on Xcode 10.3
* FBSDKCoreKit
* FBSDKShareKit
* FBSDKLoginKit
* FBSDKCoreKitSwift
* FBSDKShareKitSwift
* FBSDKLoginKitSwift

Ensure the older AND newer targets can build on the latest Xcode (11 beta 6)
// Note: Will need to clean build if you want to build the non-Swift version after building the Swift version
* FBSDKCoreKit
* FBSDKShareKit
* FBSDKLoginKit
* FBSDKCoreKitSwift
* FBSDKShareKitSwift
* FBSDKLoginKitSwift

Reviewed By: dreamolight

Differential Revision: D17164409

Pulled By: joesus

fbshipit-source-id: dbda5036487a2e626e271651b38262e56dac17ce