Adding virtual scheme for building SwiftKits (#1038)
Summary:
Thanks for proposing a pull request.

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://developers.facebook.com/opensource/cla)
- [x] describe the change (for example, what happens before the change, and after the change)

Our CI pipeline currently runs with Xcode 10.2. This is to ensure backwards compatibility for developers who are still on 10.2. Adding the Swift targets breaks the build as they are pegged to Swift 5.0. We would normally exclude these targets but Carthage is only able to build all the things.

This adds a 'composite/virtual' scheme that allows developers to build the Swift targets easily from Xcode but does not require them for CI.
Pull Request resolved: https://github.com/facebook/facebook-objc-sdk/pull/1038

Reviewed By: jingping2015

Differential Revision: D17227757

Pulled By: joesus

fbshipit-source-id: d91fc7449f829bf12506de67df601a4d69bc7c1e